Add additional models for the calculation

+++ src/Carbonara/Models/PoolHashRateDistribution/Pool.cs
@@ -3,7 +3,7 @@ namespace Carbonara.Models.PoolHashRateDistribution
     public class Pool
     {
         public string Name { get; set; }
-  	    public double Percent { get; set; }
+  	    public decimal Percent { get; set; }
   	    public string PoolType { get; set; }
     }
 }
